Improve handling of missing and misconfigured stores.
If it looks like a store is present but misconfigured (for instance,
wrong permissions) is always reported as an error.  If it looks like
the store is absent then this is only reported if --warn-store is
given, but if no configured store is present then the problems found
with all configured stores are listed.

The documentation on how to set up stores has also been clarified.