backup: report errors more uniformly
Errors creating directories and flag files are now reported the same
as failed rsync jobs.  This means that there's a greater chance
problems will show up in the backup report rather junk mail from cron.

The implementation is a bit of a kludge but should do the job for the
time being.

Problem reported by Peter Maydell.