Add a --key-file argument to rsbackup-mount
One might want to store the passphrase for a luks device in a key-file
and then use rsbackup-mount to open and mount said device for backups.
To do this, it would be handy to pass on a --key-file argument to
cryptsetup. This patch does that (and updates the usage message and
manual).

Signed-off-by: Matthew Vernon <matthew@debian.org>