Things in source tarball but not in git
See bug #809516 for the discussion that lead to this; essentially this
is a bit of a hacky workaround for the fact that the upstream tarball
doesn't match what's in git.

There are better long-term solutions, but this results in a neater
dgit history than that which would be produced by manually uploading
the built packages.