Decay pruning should preserve oldest backup in bucket
Now that there can be more than one backup per day, we need to be more
careful about how we pick the oldest backup in a bucket.

See http://www.greenend.org.uk/rjk/rsbackup/decay.pdf for the argument
that it's the oldest backup in each bucket that is preserved.

re #49