Add a NEWS.Debian.gz file documenting rejection of old logs
NEWS.Debian should be shown to people upgrading from an older version
than the version number in the NEWS entry. It is envisaged that this
NEWS will ship in rsbackup 7.0.

jessie had 1.1-4 so users upgrading should be shown the message if
they upgrade to 7.0 directly.

stretch had 3.1-3+b1 so users upgrading from stretch or later should
not be shown this message.

Signed-off-by: Matthew Vernon <matthew@debian.org>