Smarter use of --link-dest
The old policy was to link against the latest complete backup,
or the latest incomplete backup if there were no complete backups.
In the case where after a complete backup, a lot of data is added,
and that each subsequent attempt to backup fails (e.g. due to
rsync-timeout), all the work in the incomplete backup will be
fruitlessly repeated.

Now, the most recent complete backup is always used if possible,
but if there are more recent incomplete backups then the latest of
those is used as well.

fixes #77