rsbackup-snapshot-hook: swap udevadm for dmsetup
On several systems, udevadm has been observed not to give output for
some LVs for no terribly obvious reason, causing rsbackup-snapshot-hook
to refuse to work with those LVs (#72).

dmsetup also lets us canonicalise device names to work around the
original issue which caused udevadm to be invoked (#23), but is closer
to the source of truth, and thus more likely to be reliable than
udevadm.

As a side effect of this rework, VG names containing hyphens should
now be handled more convincingly.